+++ src/components/lab/TagSelector/TagSelector.tsx
@@ -85,14 +85,14 @@ export const TagSelector = forwardRef<HTMLInputElement, Props>(
 
       if (inputNode) {
         const resizeInput = () => {
-          const inputLength = inputNode.value.length
-          const isInputBlank = inputLength === 0
+          const inputNodeLength = inputNode.value.length
+          const isInputBlank = inputValue.length === 0
           const isNothingSelected = values.length === 0
           const isShowingPlaceholder = isInputBlank && isNothingSelected
 
           inputNode.style.width = isShowingPlaceholder
             ? 'auto'
-            : `${inputLength + 2}ch`
+            : `${inputNodeLength + 2}ch`
         }
 
         resizeInput()
@@ -111,12 +111,12 @@ export const TagSelector = forwardRef<HTMLInputElement, Props>(
 
     const handleKeyDown = (
       event: KeyboardEvent<HTMLInputElement>,
-      inputValue: string
+      newInputValue: string
     ) => {
       const hasSelection = values.length
       const isDeleting = event.key === 'Backspace'
 
-      if (hasSelection && !inputValue && isDeleting) {
+      if (hasSelection && !newInputValue && isDeleting) {
         handleDelete(values[values.length - 1])
       }
     }
@@ -134,10 +134,10 @@ export const TagSelector = forwardRef<HTMLInputElement, Props>(
     const handleOtherOptionSelect = (item: AutocompleteItem) => {
       const itemText = getDisplayValue!(item)
 
-      const newOption = {
+      const newOption: Item = {
         value: itemText,
         text: itemText
-      } as Item
+      }
 
       onOtherOptionSelect!(newOption)